# perf: perf yamux stream read buf