#398 use `extend_from_slice` will copy all data twice, this pr use `Vec<BytesMut>` reduces its copy from all data to just one ptr, and  Vec's memory allocation and release do not always work, because `drain` does not reclaim memory, that is, the memory allocation operation here is limited

bench master and this pr:

master branch:
```
Gnuplot not found, using plotters backend
yamux_beach/            time:   [475.10 µs 544.57 µs 618.04 µs]                         
                        change: [-5.0316% +12.625% +35.408%] (p = 0.20 > 0.05)
                        No change in performance detected.
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high mild
```
this pr:
```
Gnuplot not found, using plotters backend
yamux_beach/            time:   [393.64 µs 451.65 µs 514.75 µs]                         
                        change: [-34.776% -23.230% -8.6939%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high mild
```